requirements.txt
bcrypt==4.0.0
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyMySQL==1.0.2
Werkzeug==2.2.2
